% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iscores.R
\name{calculate_iscores}
\alias{calculate_iscores}
\title{Function that takes platforms, and calculates the minor parties' I-Scores}
\usage{
calculate_iscores(
  tibble,
  p_threshold = 0.05,
  core_threshold = 0.05,
  collapse = FALSE
)
}
\arguments{
\item{tibble}{Tibble with one row per platform, containing, at minimum (this function is designed to work with the output of `process_platform_position()`):
- `party`: The party's name (character) (this column must be unique for each platform)
- `sentence_emphasis_scores`: A list of tibbles, one per sentence in the platform (in order). Each tibble has:
   - `sentence`: The sentence (character)
   - `scores`: A tibble with the sentence's emphasis scores, containing:
       - `issue`: The issue's name (character) (every sentence in every platform must have the same issue-areas)
       - `score`: The sentence's score for that issue (numeric, summing to 100)
- `overall_emphasis_scores`: A tibble with the platform's overall emphasis scores, containing:
     - `issue`: The issue's name (character)
     - `score`: The platform's score for that issue (numeric, summing to 100)
- `position_scores`: A tibble with the platform's position-score (and standard error) for each issue-area (flagged if the Wordfish model did not converge)
- `minor_party`: Whether the party is a minor party (boolean)
- `major_party_platforms`: Only needed for minor parties. A list of lists with "before", "after", and "weight" entries, containing the name of a major party's platform before or after the minor party and the weight that should be given to the party's changes in IScore calculations.}

\item{p_threshold}{The maximum p-value for a relationship to be considered significant (0.05 by default)}

\item{core_threshold}{The minimum score a minor party must have for an issue-area for it to be considered a core issue (0.05 by default)}

\item{collapse}{Whether to remove all columns required for this function besides `party` from the final return (FALSE by default).}
}
\value{
The same tibble, only containing the minor parties, with the additional list-column `scores` containing `ie_score`, `ie_score_interpreted`, and `ip_score`.
}
\description{
Function that takes platforms, and calculates the minor parties' I-Scores
}
