% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{Function that calculates I-Score confidence intervals}
\usage{
confidence_intervals(
  tibble,
  n = 1000,
  p_threshold = 0.05,
  core_threshold = 0.05,
  exclude_nonconvergence = TRUE
)
}
\arguments{
\item{tibble}{Tibble with one row per platform, containing, at minimum (this function is designed to work with the output of `process_platform_position()`):
- `party`: The party's name (character) (this column must be unique for each platform)
- `sentence_emphasis_scores`: A list of tibbles, one per sentence in the platform (in order). Each tibble has:
   - `sentence`: The sentence (character)
   - `scores`: A tibble with the sentence's emphasis scores, containing:
       - `issue`: The issue's name (character) (every sentence in every platform must have the same issue-areas)
       - `score`: The sentence's score for that issue (numeric, summing to 1)
- `overall_emphasis_scores`: A tibble with the platform's overall emphasis scores, containing:
     - `issue`: The issue's name (character)
     - `score`: The platform's score for that issue (numeric, summing to 1)
- `position_scores`: A tibble with the platform's position-score (and standard error) for each issue-area (flagged if the Wordfish model did not converge)
- `minor_party`: Whether the party is a minor party (boolean)
- `major_party_platforms`: Only needed for minor parties. A list of lists with "before", "after", and "weight" entries, containing the name of a major party's platform before or after the minor party and the weight that should be given to the party's changes in IScore calculations.}

\item{n}{The number of bootstrap samples to take for each minor party.}

\item{p_threshold}{The maximum p-value for a relationship to be considered significant (0.05 by default).}

\item{core_threshold}{The minimum score a minor party must have for an issue-area for it to be considered a core issue (0.05 by default).}

\item{exclude_nonconvergence}{Whether to treat issues where the Wordfish model did not converge as NA when calculating Ip Scores (TRUE by default).}
}
\value{
A tibble with two rows per minor party (`upper` and `lower`), containing the upper and lower bounds of the 95% confidence interval for the party's `ie_score`, `ie_score_interpreted`, and `ip_score`.
}
\description{
Function that calculates I-Score confidence intervals
}
