% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Validate the stucture of a tibble}
\usage{
validation(tibble, set)
}
\arguments{
\item{tibble}{Tibble. The tibble input to the function calling validation().}

\item{set}{Character. The set of validators to apply ("position", "emphasis", "iscores", or "finished").}
}
\value{
Tibble of failed validation checks.
}
\description{
validation() checks the structure of an input tibble to ensure it meets the requirements for a given processing phase. The functions calling validation() should check if the returned tibble has any rows, and if so, return it to the user as an error message.
}
\keyword{internal}
