% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_score_sum.R
\name{weighted_party_scores}
\alias{weighted_party_scores}
\title{Generate IScores From A Calculation Tibble}
\usage{
weighted_party_scores(scores, weights)
}
\arguments{
\item{ie_score_tibble/ip_score_tibble}{Tibble. The calculation tibble (created during calculate_i_scores())}

\item{party_row}{Tibble. The minor party platform's row of the main tibble (the tibble input to calculate_i_scores()).}

\item{top_issues}{Character vector. The minor party platform's top issues.}
}
\value{
List. The score(s).
}
\description{
ie_score_sum/ip_score_sum() properly weights each element of an IX-Score calculation tibble to produce
}
\keyword{internal}
