% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iscores.R
\name{calculate_iscores}
\alias{calculate_iscores}
\title{Function that calculates minor parties' I-Scores}
\usage{
calculate_iscores(
  tibble,
  p_threshold = 0.05,
  core_threshold = 0.05,
  exclude_nonconvergence = TRUE,
  adjust_p_values = TRUE,
  confidence_intervals = FALSE,
  confidence_n = 1000,
  calculation_tables = FALSE
)
}
\arguments{
\item{tibble}{Tibble with one row per platform, containing, at minimum (this function is designed to work with the output of `process_platform_position()`):
- `party`: The party's name (character) (this column must be unique for each platform)
- `sentence_emphasis_scores`: A list of tibbles, one per sentence in the platform (in order). Each tibble has:
   - `sentence`: The sentence (character)
   - `scores`: A tibble with the sentence's emphasis scores, containing:
       - `issue`: The issue's name (character) (every sentence in every platform must have the same issue-areas)
       - `score`: The sentence's score for that issue (numeric, summing to 1)
- `overall_emphasis_scores`: A tibble with the platform's overall emphasis scores, containing:
     - `issue`: The issue's name (character)
     - `score`: The platform's score for that issue (numeric, summing to 1)
- `position_scores`: A tibble with the platform's position-score (and standard error) for each issue-area (flagged if the Wordfish model did not converge)
- `minor_party`: Whether the party is a minor party (boolean)
- `major_party_platforms`: Only needed for minor parties. A list of lists with "before", "after", and "weight" entries, containing the name of a major party's platform before or after the minor party and the weight that should be given to the party's changes in IScore calculations.}

\item{p_threshold}{The maximum p-value for a relationship to be considered significant (0.05 by default).}

\item{core_threshold}{The minimum score a minor party must have for an issue-area for it to be considered a core issue (0.05 by default).}

\item{exclude_nonconvergence}{Whether to treat issues where the Wordfish model did not converge as NA when calculating Ip Scores (TRUE by default).}

\item{adjust_p_values}{Whether to adjust p-values to account for the large number of comparisons (TRUE by default).}

\item{confidence_intervals}{Whether to calculate confidence intervals for the I-Scores (FALSE by default).}

\item{confidence_n}{The number of bootstrap samples to take for each minor party if `confidence_intervals` is TRUE (1000 by default).}

\item{calculation_tables}{Whether to return the tables used to calculate I-scores.}
}
\value{
A tibble, containing the minor parties, with the list-column `scores` containing `ie_score`, `ie_score_interpreted`, and `ip_score`. If `calculation_tables` is TRUE, there will be a `calculation_tables` list-column including `ie_score_table` and `ip_score_table`. If `confidence_intervals` is TRUE, there will be a `confidence_intervals` column of tibbles containing the upper and lower bounds of the 95% confidence interval for the party's IScores.
}
\description{
Function that calculates minor parties' I-Scores
}
