% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_python.R
\name{configure_python}
\alias{configure_python}
\title{Helper function to setup python tools required for processing platforms}
\usage{
configure_python(
  env_name = "iscores",
  install = TRUE,
  python_version = "latest",
  manifestoberta_model_id = "2024-1-1"
)
}
\arguments{
\item{env_name}{Name of the conda environment to setup ("iscores" by default)}

\item{install}{Whether to install missing packages (TRUE by default)}

\item{python_version}{Version of Python to use if a new environment must be created ("latest" by default)}

\item{manifestoberta_model_id}{Version of the ManifestoBERTA model to use ("2024-1-1" by default)}
}
\value{
Returns invisible(TRUE) when setup is successful, otherwise it provides a descriptive error
}
\description{
Helper function to setup python tools required for processing platforms
}
