% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_platform_emphasis.R
\name{process_platform_emphasis}
\alias{process_platform_emphasis}
\title{Function that takes platforms, splits them into sentences, and calculates issue-area emphasis scores}
\usage{
process_platform_emphasis(tibble, cleaning = TRUE)
}
\arguments{
\item{tibble}{Tibble with one row per platform, containing, at minimum:
- `text`: A character column with the full text of each platform}

\item{cleaning}{Whether to apply basic text-cleaning before processing platforms (TRUE by default)}
}
\value{
The same tibble with two rm list columns (if a platform cannot be processed due to a lack of text, the function will return an empty list for that platform):
 - `sentence_emphasis_scores`: A list of tibbles, one per sentence in the platform (in order). Each tibble has:
    - `sentence`: The sentence (character)
    - `scores`: A tibble with the sentence's emphasis scores, containing:
        - `issue`: The issue's name (character)
        - `score`: The sentence's score for that issue (numeric, summing to 1)
 - `overall_emphasis_scores`: A tibble with the platform's overall emphasis scores, containing:
    - `issue`: The issue's name (character)
    - `score`: The platform's score for that issue (numeric, summing to 1)
}
\description{
Function that takes platforms, splits them into sentences, and calculates issue-area emphasis scores
}
